- a big change to the mesh rendering. the four large tiles are now decimated into smaller tiles so that there is just one texture per tile (512 small textures). The idea is that moving small textures takes less time per texture than moving large textures... Also, this could eliminate some "bleed" of one tile into another with some forms of texture filtering - some changes to the internals of the map/mesh - come changes to the cartman project files so that they will compile under vs10 - a "fix" to the toupper "function" so that it works equivalently in vs7, vs10, and hopefully gcc. Long ago it was implemented as an array, but in vs 10 it is defined as a function taking an int and returning an int. This caused a lot of useless warnings, but also some possible problems. - added some guard code to vfs_getc(), which actually returns an int. Things like end-of-file give numbers outside the range of a char, and they must be handled as ints. - fixed some notes to follow the doxygen format - fixed some potential problems dealing with the way that the character skin information being stored. Depending on how the skin-override or spelleffect-override expansions cap file, the value could be -1 for the skin at startup and this has to be preserved in several places. - some minor type casting to remove silly warnings - some removal of += and -= from ints of limited size or unsigned ints because of unpredictable behavior - a minor fix to the GiveTargetMoney function so that no one can have negative money and no one can go over their limit - fixed a confusing variable name and an incorrect varable type in the Zefz's implementation of the player menus - forced the renderlist to sort elements by tile type and distance so that close tiles are rendered first (possibly blocking other tiles and removing them from ever being rendered, and hopefully saving time) - removed a potential problem with out-of-range characters being stored in the key buffer when in terminal-mode

+++ game/camera.c
@@ -945,7 +945,7 @@ bool_t camera_reset_target( camera_t * pcam, const ego_mesh_t * pmesh, const CHR
     ///          a "change of scene". With the new velocity-tracking of the camera, this would include
     ///          things like character respawns, adding new players, etc.
 
-    int turn_mode_save, move_mode_save;
+    Uint8 turn_mode_save, move_mode_save;
 
     if ( NULL == pcam ) return bfalse;
